finished promotions

+++ app-order/src/main/java/uz/pdp/apporder/entity/Order.java
@@ -47,7 +47,7 @@ public class Order {
 
     private Float deliverySum;
 
-    private Double overAllSum;
+    private Float overAllSum;
 
     private Integer number;
